Added user authorization to delete_todo method

+++ fastapi/todo/main.py
@@ -80,8 +80,11 @@ async def update_todo(todo_id: int, todo: Todo, user: dict = Depends(get_current
     
     
 @app.delete("/{todo_id}")
-async def delete_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    if user is None:
+        raise get_user_exception() 
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model:
         db.query(Todos).filter(Todos.id == todo_id).delete()
         db.commit()
